+++ admin/src/assets/styles/components/_Panel.scss
@@ -218,6 +218,10 @@
 		padding-right: 1.5em;
 		overflow-y: scroll;
 
+		&.no-scrollbar {
+			overflow-y: visible;
+		}
+
 		@media (min-height: 31.25em) {
 			max-height: 50vh;
 		}
